requirements.txt
requests
selenium
pandas
html5lib
webdriver-manager
opencv-python
numpy
yfinance
ta
twilio
python-dotenv
google-genai
Pillow